Fix CI: JS linking (di-cats IO, optics Array) + 2.13 doc snippets
- di-cats: `cats.effect.IO.unsafeRunSync()` is JVM-only (IO cannot block on Scala.js).
  Moved the two IO F-agnosticism tests from the shared `ResourceWiringSpec` into a JVM-only
  `ResourceWiringJvmSpec` (`src/test/scalajvm`); cross-platform F-abstraction stays covered by
  the `SyncIO` tests.
- optics: `arrayQuicklensFunctor` used `java.lang.reflect.Array`, which does not link on
  Scala.js. Since `.each` keeps the element type (`A => A`), rebuild via `Array.clone()` +
  in-place update — cross-platform (JVM/JS/Native).
- docs: `optics.md` used `import hearth.kindlings.optics.*` (Scala-3-only); the snippet runner
  compiles on plain 2.13 (no `-Xsource:3`), so switched all snippet imports to `._` (matching
  di.md / mock.md).

Verified: di-cats JS 32 / JVM 34, optics JS+JVM 69 on both Scala versions.
