+++ docs/user-guide/optics.md
@@ -31,7 +31,7 @@ nested `.copy` chain by hand:
 ```scala
 //> using dep com.kubuszok::kindlings-optics:{{ kindlings_version() }}
 
-import hearth.kindlings.optics.*
+import hearth.kindlings.optics._
 
 final case class Street(name: String)
 final case class Address(street: Street)
@@ -62,7 +62,7 @@ println(updated.address.street.name)
 ```scala
 //> using dep com.kubuszok::kindlings-optics:{{ kindlings_version() }}
 
-import hearth.kindlings.optics.*
+import hearth.kindlings.optics._
 
 final case class Person(name: String, age: Int)
 val p = Person("Anna", 30)
@@ -81,7 +81,7 @@ by providing a `QuicklensFunctor[F]`.)
 ```scala
 //> using dep com.kubuszok::kindlings-optics:{{ kindlings_version() }}
 
-import hearth.kindlings.optics.*
+import hearth.kindlings.optics._
 import hearth.kindlings.optics.syntax.*
 
 final case class Team(members: List[String])
@@ -101,7 +101,7 @@ inserts the default when absent. The same three names also work on a `Map` (keye
 ```scala
 //> using dep com.kubuszok::kindlings-optics:{{ kindlings_version() }}
 
-import hearth.kindlings.optics.*
+import hearth.kindlings.optics._
 import hearth.kindlings.optics.syntax.*
 
 final case class Scores(values: List[Int])
@@ -120,7 +120,7 @@ the other branch.
 ```scala
 //> using dep com.kubuszok::kindlings-optics:{{ kindlings_version() }}
 
-import hearth.kindlings.optics.*
+import hearth.kindlings.optics._
 import hearth.kindlings.optics.syntax.*
 
 final case class Result(value: Either[String, Int])
@@ -139,7 +139,7 @@ non-exhaustive, 2-case match generated by the macro).
 ```scala
 //> using dep com.kubuszok::kindlings-optics:{{ kindlings_version() }}
 
-import hearth.kindlings.optics.*
+import hearth.kindlings.optics._
 import hearth.kindlings.optics.syntax.*
 
 sealed trait Animal
@@ -162,7 +162,7 @@ rebuilds it with a **single** `.copy` carrying both updated leaves (true prefix 
 ```scala
 //> using dep com.kubuszok::kindlings-optics:{{ kindlings_version() }}
 
-import hearth.kindlings.optics.*
+import hearth.kindlings.optics._
 
 final case class Names(first: String, second: String, third: String)
 val n = Names("ann", "bob", "cid")
@@ -177,7 +177,7 @@ A `.each` path can be mixed with a plain field path; the modification is applied
 ```scala
 //> using dep com.kubuszok::kindlings-optics:{{ kindlings_version() }}
 
-import hearth.kindlings.optics.*
+import hearth.kindlings.optics._
 
 final case class Box(xs: List[String], name: String)
 val box = Box(List("ann", "bob"), "label")
@@ -195,7 +195,7 @@ terminal operations return a `T => T` function you can apply to many values:
 ```scala
 //> using dep com.kubuszok::kindlings-optics:{{ kindlings_version() }}
 
-import hearth.kindlings.optics.*
+import hearth.kindlings.optics._
 
 final case class Person(name: String, age: Int)
 
@@ -218,7 +218,7 @@ Two lenses compose with `andThenModify`: a `PathLazyModify[T, U]` followed by a 
 ```scala
 //> using dep com.kubuszok::kindlings-optics:{{ kindlings_version() }}
 
-import hearth.kindlings.optics.*
+import hearth.kindlings.optics._
 
 final case class Address(city: String, street: String)
 final case class Person(name: String, address: Address)
